+++ sdk/ai/azure-ai-projects/samples/agents/async_samples/sample_agents_vector_store_file_search_async.py
@@ -20,7 +20,11 @@ import asyncio
 import os
 
 from azure.ai.projects.aio import AIProjectClient
-from azure.ai.projects.models import FileSearchTool, FilePurpose
+from azure.ai.projects.models import (
+    FileSearchTool,
+    FilePurpose,
+    MessageTextContent
+)
 from azure.identity.aio import DefaultAzureCredential
 
 
@@ -74,10 +78,12 @@ async def main():
             messages = await project_client.agents.list_messages(thread_id=thread.id)
             
             for message in reversed(messages.data):
+                if not isinstance(message.content[0], MessageTextContent):
+                    continue
                 # To remove characters, which are not correctly handled by print, we will encode the message
                 # and then decode it again. 
                 clean_message = message.content[0].text.value.encode('ascii', 'ignore').decode('utf-8')
-                print(f"Role: {message.role.value}  Message: {clean_message}")
+                print(f"Role: {message.role}  Message: {clean_message}")
 
 
 if __name__ == "__main__":
